refactor: move canonical url to separate component
