+++ src/components/pages/contact-sales/hero/contact-form/contact-form.jsx
@@ -9,7 +9,7 @@ import * as yup from 'yup';
 import Button from 'components/shared/button';
 import Field from 'components/shared/field';
 import Link from 'components/shared/link';
-import { HUBSPOT_CONTACT_SALES_FORM_ID, FORM_STATES } from 'constants/forms';
+import { FORM_STATES, HUBSPOT_CONTACT_SALES_FORM_ID } from 'constants/forms';
 import { doNowOrAfterSomeTime, sendHubspotFormData } from 'utils/forms';
 
 const schema = yup
@@ -22,6 +22,8 @@ const schema = yup
     message: yup.string().required('Message is a required field'),
   })
   .required();
+
+console.log(schema);
 
 const ContactForm = ({ formState, setFormState }) => {
   const {
