+++ src/components/shared/seo/canonical-url/index.js
@@ -1,0 +1,3 @@
+import CanonicalUrl from './canonical-url';
+
+export default CanonicalUrl;
