+++ src/components/shared/seo/seo.jsx
@@ -1,6 +1,4 @@
-'use client';
-
-import { usePathname } from 'next/navigation';
+import CanonicalUrl from 'components/shared/seo/canonical';
 import PropTypes from 'prop-types';
 
 const SITE_URL = process.env.NEXT_PUBLIC_DEFAULT_SITE_URL || 'localhost:3000';
@@ -23,8 +21,6 @@ const SEO = ({
   imagePath = defaultImagePath,
 }) => {
   const currentImagePath = imagePath.startsWith('http') ? imagePath : SITE_URL + imagePath;
-  const pathname = usePathname();
-  const canonicalUrl = process.env.NEXT_PUBLIC_DEFAULT_SITE_URL + pathname;
 
   return (
     <>
@@ -34,12 +30,11 @@ const SEO = ({
         content="width=device-width, initial-scale=1, shrink-to-fit=no, viewport-fit=cover"
       />
       <title>{title}</title>
-      <link rel="canonical" href={canonicalUrl} />
+      <CanonicalUrl />
       <meta name="description" content={description} key="desc" />
       <meta property="og:title" content={title} />
       <meta property="og:description" content={description} />
       <meta property="og:image" content={currentImagePath} />
-      <meta property="og:url" content={canonicalUrl} />
       <meta property="twitter:title" content={title} />
       <meta property="twitter:description" content={description} />
       <meta name="twitter:card" content="summary_large_image" />
